Add PhotoDataFragment
DetailsWrapperFragment is a Fragment that shows full description of a
photo - its title, date, tags and url. Currently instead of tags url is
being shown - that needs fixing in future commits.

Add PhotoSimilarFragment

PhotoSimilarFragment is a Fragment that shows up to 6 images similar to
the source image. Currently instead of similar images a placeholder TV
is being shown.

Add DetailsWrapperFragment

DetailsWrapperFragment is a Fragment that wraps up PhotoDataFramgent and
PhotoSimilarFragment. It is necessary for TabLayout in
PhotoDetailsActivity.

Modify PhotoDetailsActivity - Show proper titles of fragments on top
bar.